+++ sigma/modules/moderation/messages/textmute.py
@@ -19,7 +19,6 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 import arrow
 import discord
 
-from sigma.core.mechanics.database import Database
 from sigma.core.mechanics.incident import get_incident_core
 from sigma.core.utilities.data_processing import convert_to_seconds, user_avatar
 from sigma.core.utilities.event_logging import log_event
@@ -29,15 +28,14 @@ from sigma.core.utilities.permission_processing import hierarchy_permit
 
 def generate_log_embed(message, target, reason):
     """
-
     :param message:
-    :type message:
+    :type message: discord.Message
     :param target:
-    :type target:
+    :type target: discord.Member
     :param reason:
-    :type reason:
+    :type reason: str
     :return:
-    :rtype:
+    :rtype: discord.Embed
     """
     log_embed = discord.Embed(color=0x696969, timestamp=arrow.utcnow().datetime)
     log_embed.set_author(name='A Member Has Been Muted', icon_url=user_avatar(target))
@@ -52,19 +50,19 @@ def generate_log_embed(message, target, reason):
     return log_embed
 
 
-async def make_incident(db: Database, gld: discord.Guild, ath: discord.Member, trg: discord.Member, reason: str):
+async def make_incident(db, gld, ath, trg, reason):
     """
 
     :param db:
-    :type db:
+    :type db: sigma.core.mechanics.database.Database
     :param gld:
-    :type gld:
+    :type gld: discord.Guild
     :param ath:
-    :type ath:
+    :type ath: discord.Member
     :param trg:
-    :type trg:
+    :type trg: discord.Member
     :param reason:
-    :type reason:
+    :type reason: str
     """
     icore = get_incident_core(db)
     inc = icore.generate('textmute')
@@ -113,18 +111,18 @@ async def textmute(cmd, pld):
                     if mute_list is None:
                         mute_list = []
                     if target.id in mute_list:
-                        response = error(f'{target.display_name} is already text muted.')
+                        response = error(f'{target.display_name} is already text-muted.')
                     else:
                         mute_list.append(target.id)
                         await cmd.db.set_guild_settings(pld.msg.guild.id, 'muted_users', mute_list)
-                        response = ok(f'{target.display_name} has been text muted.')
+                        response = ok(f'{target.display_name} has been text-muted.')
                         rarg = pld.args[1:-1] if timed else pld.args[1:] if pld.args[1:] else None
                         reason = ' '.join(rarg) if rarg else None
                         await make_incident(cmd.db, pld.msg.guild, pld.msg.author, target, reason)
                         log_embed = generate_log_embed(pld.msg, target, reason)
                         await log_event(cmd.bot, pld.settings, log_embed, 'log_mutes')
                         guild_icon = str(pld.msg.guild.icon_url) if pld.msg.guild.icon_url else discord.Embed.Empty
-                        to_target_title = '🔇 You have been text muted.'
+                        to_target_title = '🔇 You have been text-muted.'
                         to_target = discord.Embed(color=0x696969)
                         to_target.add_field(name=to_target_title, value=f'Reason: {reason}')
                         to_target.set_footer(text=f'On: {pld.msg.guild.name}', icon_url=guild_icon)